requirements.txt
alabaster==0.7.11
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
argcomplete==1.9.4
asn1crypto==0.24.0
astroid==2.0.4
astropy==3.0.4
atomicwrites==1.1.5
attrs==18.1.0
Automat==0.7.0
autopep8==1.4
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bleach==2.1.4
bokeh==0.13.0
boto==2.49.0
Bottleneck==1.2.1
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
chest==0.2.3
click==6.7
cloudpickle==0.5.3
colorama==0.3.9
configobj==5.0.6
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.3
cryptography-vectors==2.3
cycler==0.10.0
Cython==0.28.5
cytoolz==0.9.0.1
dask==0.18.2
decorator==4.3.0
dill==0.2.8.2
distributed==1.22.1
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.6
Flask-PyMongo==2.1.0
gevent==1.3.6
gmpy2==2.0.8
greenlet==0.4.14
gunicorn==19.9.0
h5py==2.8.0
HeapDict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
icc-rt==2019.0
idna==2.7
imageio==2.3.0
imagesize==1.0.0
incremental==17.5.0
intel-numpy==1.14.3
intel-openmp==2019.0
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.24.0
locket==0.2.0
lxml==4.2.4
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.3
mkl==2019.0
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
multipledispatch==0.6.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.6.0
numba==0.39.0
numexpr==2.6.7
numpy==1.15.0
numpydoc==0.8.0
olefile==0.45.1
openpyxl==2.5.5
packaging==17.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pkginfo==1.4.2
pluggy==0.7.1
ply==3.11
prometheus-client==0.3.1
prompt-toolkit==1.0.15
psutil==5.4.7
ptyprocess==0.6.0
py==1.5.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.2.0
pylint==2.1.1
pymongo==3.7.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
PyQt5==5.9.2
PySocks==1.6.8
pytest==3.7.2
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.0
QtPy==1.4.2
redis==2.10.6
requests==2.19.1
rope==0.11.0
ruamel.yaml==0.15.46
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip==4.19.8
six==1.11.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
sortedcontainers==2.0.4
Sphinx==1.7.7
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.10
statsmodels==0.9.0
sympy==1.2
tables==3.4.4
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.1
traitlets==4.3.2
Twisted==17.5.0
typed-ast==1.1.0
unicodecsv==0.14.1
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.0
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.7
xlwings==0.11.8
xlwt==1.2.0
zict==0.1.3
zope.interface==4.5.0
